fix bug that cann't recognize ONBUILD in schemaV1 image
fix bug #3052

Original `convertSchema1ConfigMeta` fuction using
the wrong `ocispec.Image` type to unmarshal schemaV1
image config, Which miss ONBUILD field, the correct
`Image` type was defined in `frontend/dockerfile/dockerfile2llb/image.go`,
I move its definition to `exporter/containerimage/image`,
then both package `util/imageutil` and `frontend/dockerfile`
could use the correct `Image` type.

(cherry picked from commit 9dd3dfdf9e9692c6a754491b2db95132761ecca9)
